fixes: yt video recommendation
